fix updateTradeArea infinite loop + warnings

I also tried renaming a previous commit... that didn't work well

but mostly everything works :D

+++ src/Utils.h
@@ -26,7 +26,7 @@ std::string getLine(int minChars)
     {
         std::string line;
         std::getline(std::cin, line);
-        if (line.length() >= minChars)
+        if (line.length() >= (size_t) minChars)
             return line;
         else
             std::cout << "Please enter at least " << minChars << " characters" << std::endl;
